% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsVec.R
\name{tsVec}
\alias{tsVec}
\title{A Time Vector Function}
\usage{
tsVec(x, y, BY = c("month", "week", "day", "hour"), TZ = "EST")
}
\arguments{
\item{Starting}{Date, Ending Date, Time Split, Time Zone (defaults to EST)}
}
\description{
This function allows for the creation of a vector between two dates for TS data handling.
}
\examples{
tsVec()

}
\keyword{time-series}

